--- a/src/extension.rs
+++ b/src/extension.rs
@@ -5,7 +5,6 @@ use std::future::Future;
 use bevy::prelude::{Commands, EntityWorldMut, World};
 
 use crate::prelude::Reactor;
-use crate::reactor::Initialized;
 use crate::task::ReactiveTask;
 use crate::world_ptr::WorldPtr;
 
@@ -28,10 +27,8 @@ impl<'w, Fun, Fut> ScheduleReactor<'w, Fun, Fut, EntityWorldMut<'w>> for World
     fn spawn_initialized_reactor(&'w mut self, f: Fun) -> EntityWorldMut<'w> {
         let mut reactor = Reactor::schedule(f);
         reactor.run_sync(WorldPtr::new(self));
-        self.spawn((
-            Initialized,
-            reactor
-        ))
+        reactor.initialized = true;
+        self.spawn(reactor)
     }
 }