--- a/src/lib.rs
+++ b/src/lib.rs
@@ -13,9 +13,9 @@
 use bevy::app::{App, Last, MainScheduleOrder, Plugin, PostStartup};
 use bevy::ecs::schedule::ScheduleLabel;
 use bevy::hierarchy::DespawnRecursiveExt;
-use bevy::prelude::{Entity, QueryState, Without, World};
+use bevy::prelude::{Entity, QueryState, World};
 
-use crate::reactor::{Initialized, Reactor};
+use crate::reactor::Reactor;
 use crate::world_ptr::WorldPtr;
 
 pub mod extension;
@@ -35,8 +35,8 @@ pub mod prelude {
             Record,
             Redo,
             RedoAction,
-            Track,
             Rollback,
+            Track,
             Undo,
             UndoRedoInProgress,
         },
@@ -86,46 +86,34 @@ struct RunReactor;
 
 fn initialize_reactors(
     world: &mut World,
-    reactors: &mut QueryState<(Entity, &mut Reactor), Without<Initialized>>,
+    reactors: &mut QueryState<&mut Reactor>,
 ) {
     let world_ptr = WorldPtr::new(world);
-    for (entity, mut reactor) in reactors.iter_mut(world) {
-        world_ptr.as_mut().entity_mut(entity).insert(Initialized);
+    for mut reactor in reactors.iter_mut(world) {
+        if reactor.initialized {
+            continue;
+        }
+        reactor.initialized = true;
         reactor.run_sync(world_ptr);
     }
 }
 
 fn run_reactors(
     world: &mut World,
-    reactors: &mut QueryState<(Entity, &mut Reactor, Option<&Initialized>)>,
+    reactors: &mut QueryState<(Entity, &mut Reactor)>,
 ) {
-    enum Status {
-        Finished,
-        Initialized,
-    }
-
     let world_ptr = WorldPtr::new(world);
-    let mut entities = Vec::with_capacity(reactors.iter(world).len());
-    for (entity, mut reactor, initialized) in reactors.iter_mut(world) {
-        if initialized.is_none() {
+    for (entity, mut reactor) in reactors.iter_mut(world) {
+        if !reactor.initialized {
             if reactor.run_sync(world_ptr) || reactor.run_sync(world_ptr) {
-                entities.push((entity, Status::Finished));
+                world_ptr.as_mut().entity_mut(entity).despawn_recursive();
+                // entities.push((entity, Status::Finished));
             } else {
-                entities.push((entity, Status::Initialized));
+                reactor.initialized = true;
             }
         } else if reactor.run_sync(world_ptr) {
-            entities.push((entity, Status::Finished));
-        }
-    }
-
-    for (entity, status) in entities {
-        match status {
-            Status::Finished => {
-                world.entity_mut(entity).despawn_recursive();
-            }
-            Status::Initialized => {
-                world.entity_mut(entity).insert(Initialized);
-            }
+            world_ptr.as_mut().entity_mut(entity).despawn_recursive();
+            // entities.push((entity, Status::Finished));
         }
     }
 }