--- a/src/runner.rs
+++ b/src/runner.rs
@@ -81,9 +81,7 @@ pub(crate) fn initialize_schedule(schedules: &mut Schedules, schedule_label: Int
 fn run_runners<L: Send + Sync + 'static>(world: &mut World) {
     if let Some(mut runners) = world.remove_non_send_resource::<BoxedRunners<L>>() {
         runners.0.retain_mut(|(runner, token)| {
-            if token.finished_reactor() {
-                false
-            } else if token.is_cancellation_requested() {
+            if token.is_cancellation_requested() {
                 token.call_cancel_handles(world);
                 false
             } else {