I optimized `Reactor` related processing.

```
before: with_flurx count: 10000 time:   [96.151 ms 97.769 ms 99.598 ms]
after : with_flurx count: 10000 time:   [92.705 ms 94.415 ms 96.343 ms]
```